添加项目用户列表

+++ image-server/src/main/java/com/clsaa/dop/server/image/feign/harborfeign/ProjectFeign.java
@@ -179,7 +179,7 @@ public interface ProjectFeign {
      * @param auth harbor的权限参数
      */
     @PostMapping(value = "/projects/{project_id}/members")
-    void  projectsProjectIdMembersPost(@PathVariable("project_id") Long projectId,
+    void  projectsProjectIdMembersPost(@PathVariable("project_id") Integer projectId,
                                        @RequestBody ProjectMember projectMember,
                                        @RequestHeader(value = "Authorization") String auth);
     /**
